fix: issue with `previous()` function for consumables and tools.
Let me describe the bug:
If you roll for 1 time, then exclude tool or cons. from re-roll. Then hit roll and try to click on previous tool/cons.
It will show error in console. I know that it's not a big problem, but I'm glad to fix it.
I'd not searched for the reason of `null` in tool slot in storage, but inserted check for its value to return early if it's needed.